feat: Add category formatting and MongoDB document creation for project categories
- Implemented formatCategories function to validate and format category data.
- Added createCategoryDocuments function to generate MongoDB documents for categories.
- Refactored processCategories function to utilize new formatting and document creation functions.
- Introduced formatTaskDocument function for consistent task data formatting.
- Created processProjectAsTask function to validate and publish project templates.
- Added user mapping helper functions for managing user-program relationships.
- Created certificate helper functions for handling certificate template operations.
- Removed unused code and optimized existing functions for better readability and maintainability.
