+++ src/consumption/elevate.js
@@ -17,37 +17,19 @@ const targetingHelpers = require('@helpers/targetingCriteria')
 const { Op } = require('sequelize')
 const rolloutQueries = require('@database/queries/rollouts')
 const { ObjectId } = require('mongodb')
-const fs = require('fs')
-const axios = require('axios')
-const cheerio = require('cheerio')
-const path = require('path')
 const certificateBaseTemplateQueries = require('@database/queries/certificateBaseTemplate')
 const userRequests = require('@requests/user')
-const interfaceRequests = require('@requests/interface')
 let projectsMongoConnection = null
 let mongoConnection = null
 let scopeKeys = {}
-let socketInUse = false
 const resourceQueries = require('@database/queries/resources')
 const projectDTO = require('@consumption/dtos/elevate/project')
 const programDTO = require('@consumption/dtos/elevate/program')
 const solutionDTO = require('@consumption/dtos/elevate/solution')
 const targetingHelper = require('@consumption/helpers/elevate/targeting')
-
-// Define the mongoDb collection names used
-const COLLECTIONS_MAP = new Map(
-	Object.entries({
-		CATEGORIES: 'projectCategories',
-		TEMPLATES: 'projectTemplates',
-		TASKS: 'projectTemplateTasks',
-		PROGRAMS: 'programs',
-		SOLUTIONS: 'solutions',
-		CERTIFICATE_TEMPLATE: 'certificateTemplates',
-		CERTIFICATE_BASE_TEMPLATE: 'certificateBaseTemplates',
-		USER_EXTENSIONS: 'userExtensions',
-		ORGANIZATION_EXTENSION: 'organizationExtension',
-	})
-)
+const userMappingHelper = require('@consumption/helpers/elevate/userMapping')
+const certificateHelper = require('@consumption/helpers/elevate/certificate')
+const { COLLECTIONS_MAP } = require('@consumption/constants/elevate/common')
 
 /**
  * To connect with the mongoDB with the given url
@@ -74,21 +56,16 @@ const connectMongo = async (url) => {
  */
 async function processCategories(categories, orgCode, tenantCode) {
 	try {
+		// Format categories using DTO
+		const formattedCategoriesResponse = projectDTO.formatCategories(categories)
+		if (!formattedCategoriesResponse.success) {
+			throw new Error(formattedCategoriesResponse.error)
+		}
+
+		const formattedCategories = formattedCategoriesResponse.data
+
 		// Fetch a specific collection
 		const categoriesCollection = projectsMongoConnection.collection(COLLECTIONS_MAP.get('CATEGORIES'))
-
-		// Format categories
-		const formattedCategories = categories.map((category) => {
-			if (!category.label || !category.value) {
-				throw new Error('EACH_CATEGORY_MUST_BE_LABEL_AND_VALUE')
-			}
-			return {
-				label: category.label,
-				value: category.value,
-				formattedName: utils.formatToTitleCase(category.value),
-				externalId: category.value.replace(/_/g, '').toLowerCase(),
-			}
-		})
 
 		// Fetch existing categories by externalId
 		let existingCategories = []
@@ -101,25 +78,10 @@ async function processCategories(categories, orgCode, tenantCode) {
 
 		const existingExternalIds = [...new Set(existingCategories.map((cat) => cat.externalId))]
 
-		// Filter out categories that already exist
-		const newCategories = formattedCategories
-			.filter((cat) => !existingExternalIds.includes(cat.externalId))
-			.map(({ formattedName, externalId, label }) => ({
-				createdBy: 'SYSTEM',
-				updatedBy: 'SYSTEM',
-				isDeleted: false,
-				isVisible: true,
-				status: 'active',
-				icon: '',
-				noOfProjects: 0,
-				name: formattedName,
-				externalId: externalId,
-				label: label,
-				tenantId: tenantCode,
-				orgId: orgCode,
-				createdAt: new Date(),
-				updatedAt: new Date(),
-			}))
+		// Filter out categories that already exist and create MongoDB documents
+		const categoriesToInsert = formattedCategories.filter((cat) => !existingExternalIds.includes(cat.externalId))
+
+		const newCategories = projectDTO.createCategoryDocuments(categoriesToInsert, orgCode, tenantCode)
 
 		// Insert only new categories
 		if (newCategories.length > 0) {
@@ -161,6 +123,80 @@ async function processCategories(categories, orgCode, tenantCode) {
 }
 
 /**
+ * Process project as a task - validate, publish if needed, and get template details
+ * @name processProjectAsTask
+ * @param {Object} task - Task data
+ * @param {String} tenantCode - Tenant code
+ * @param {String} organizationCode - Organization code
+ * @returns {Object} - Response with projectTemplateDetails or error
+ */
+async function processProjectAsTask(task, tenantCode, organizationCode) {
+	try {
+		const projectTemplatesCollection = projectsMongoConnection.collection(COLLECTIONS_MAP.get('TEMPLATES'))
+
+		// Validate project exists and is published
+		const validateProject = await resourceQueries.findOne(
+			{
+				id: task.project_id,
+				type: common.PROJECT,
+				status: common.RESOURCE_STATUS_PUBLISHED,
+				tenant_code: tenantCode,
+			},
+			[]
+		)
+
+		if (!validateProject?.id) {
+			throw new Error(`Project not found or not published for task project_id: ${task.project_id}`)
+		}
+
+		// If task project not published, publish it
+		if (!validateProject?.published_id) {
+			let publishedProject = await publishProjectTemplates({
+				id: task.project_id,
+				tenant_code: tenantCode,
+				organization_code: organizationCode,
+			})
+			if (!publishedProject.success || !publishedProject?.templateId) {
+				throw new Error(`Task Project template publish failed: ${validateProject.project_id}`)
+			}
+			validateProject.published_id = publishedProject.templateId
+		}
+
+		// Convert published_id to ObjectId
+		const publishedObjectId = new ObjectId(validateProject.published_id)
+
+		// Get projectTemplate details
+		let projectTemplates = await projectTemplatesCollection.findOne({
+			_id: publishedObjectId,
+			tenantId: tenantCode,
+		})
+
+		if (!projectTemplates) {
+			throw new Error(`Project template not found for published_id: ${validateProject.published_id}`)
+		}
+
+		// Return project template details and type
+		return {
+			success: true,
+			projectTemplateDetails: {
+				_id: projectTemplates._id,
+				type: projectTemplates.type ?? common.SOLUTIONS_TYPE.project,
+				entityType: projectTemplates.entityType,
+				isReusable: projectTemplates.isReusable,
+				externalId: projectTemplates.externalId,
+			},
+			type: projectTemplates.type ?? common.SOLUTIONS_TYPE.project,
+		}
+	} catch (error) {
+		console.error('Error in processProjectAsTask:', error.message)
+		return {
+			success: false,
+			error: error.message,
+		}
+	}
+}
+
+/**
  * Create Task
  * @name createTasks
  * @param {Object} tasks - task data
@@ -177,26 +213,16 @@ async function createTasks(tasks, templateId, templateExternalId, parentId = nul
 		const externalIds = []
 
 		for (const task of tasks) {
-			// Format the task data
-			let taskData = {
-				name: task.name,
-				description: task.name,
-				externalId: utils.generateExternalId(task.name),
-				type: task.type,
-				isDeleted: false,
-				isDeletable: !task.is_mandatory,
-				sequenceNumber: task.sequence_no,
-				projectTemplateId: templateId,
-				projectTemplateExternalId: templateExternalId,
-				hasSubTasks: task.children?.length > 0,
-				learningResources: utils.convertResources(task.learning_resources || []),
+			// Format the task data using DTO
+			let taskData = projectDTO.formatTaskDocument(
+				task,
+				templateId,
+				templateExternalId,
 				parentId,
-				deleted: false,
-				orgId: organizationCode,
-				tenantId: tenantCode,
-				createdAt: new Date(),
-				updatedAt: new Date(),
-			}
+				organizationCode,
+				tenantCode
+			)
+
 			//if task type reflection add link and buttonLabel in task metaInformation
 			if (task.type === common.TASK_TYPE_REFLECTION) {
 				taskData.metaInformation = {
@@ -207,54 +233,14 @@ async function createTasks(tasks, templateId, templateExternalId, parentId = nul
 
 			//if task type improvementProject add projectTemplateDetails
 			if (task.type === common.TASK_TYPE_PROJECT) {
-				const projectTemplatesCollection = projectsMongoConnection.collection(COLLECTIONS_MAP.get('TEMPLATES'))
-				const validateProject = await resourceQueries.findOne(
-					{
-						id: task.project_id,
-						type: common.PROJECT,
-						status: common.RESOURCE_STATUS_PUBLISHED,
-						tenant_code: tenantCode,
-					},
-					[]
-				)
-				if (!validateProject?.id) {
-					throw new Error(`Project not found or not published for task project_id: ${task.project_id}`)
-				}
-				//if task project not published then will publish
-				if (!validateProject?.published_id) {
-					let publishedProject = await publishProjectTemplates({
-						id: task.project_id,
-						tenant_code: tenantCode,
-						organization_code: organizationCode,
-					})
-					if (!publishedProject.success || !publishedProject?.templateId) {
-						throw new Error(`Task Project template publish failed: ${validateProject.project_id}`)
-					}
-					validateProject.published_id = publishedProject.templateId
+				const projectResult = await processProjectAsTask(task, tenantCode, organizationCode)
+
+				if (!projectResult.success || !projectResult?.projectTemplateDetails) {
+					throw new Error(`Failed to process project as task: ${projectResult.error}`)
 				}
 
-				// Convert published_id to ObjectId
-				const publishedObjectId = new ObjectId(validateProject.published_id)
-
-				// get projectTemplate details
-				let projectTemplates = await projectTemplatesCollection.findOne({
-					_id: publishedObjectId,
-					tenantId: tenantCode,
-				})
-
-				if (!projectTemplates) {
-					throw new Error(`Project template not found for published_id: ${validateProject.published_id}`)
-				}
-
-				taskData.projectTemplateDetails = {
-					_id: projectTemplates._id,
-					type: projectTemplates.type ?? common.SOLUTIONS_TYPE.project,
-					entityType: projectTemplates.entityType,
-					isReusable: projectTemplates.isReusable,
-					externalId: projectTemplates.externalId,
-				}
-
-				taskData.type = projectTemplates.type ?? common.SOLUTIONS_TYPE.project
+				taskData.projectTemplateDetails = projectResult?.projectTemplateDetails
+				taskData.type = projectResult?.type
 			}
 
 			// Create the task
@@ -349,6 +335,7 @@ const publishProjectTemplates = function (templateData) {
 			projectsMongoConnection = projectsMongoConnection
 				? projectsMongoConnection
 				: await connectMongo(projectsMongoDBUrl)
+
 			// Fetch Org Policies
 			const orgPolicies = await fetchOrgPolicies(
 				templateData.organization_code,
@@ -586,272 +573,6 @@ async function createProgram(programTemplate) {
 }
 
 /**
- *  Function to fetch data information from cloud using downloadable Url
- * @method
- * @name getBaseTemplate
- * @param {String} templateUrl - cloud path to download
- */
-async function getBaseTemplate(templateUrl) {
-	try {
-		// Mark socket as engaged
-		socketInUse = true
-		const response = await axios.get(templateUrl)
-		// Mark socket as free
-		socketInUse = false
-		if (response.status === 200) {
-			return {
-				success: true,
-				result: response.data,
-			}
-		} else {
-			throw new Error(`Unexpected response status: ${response.status}`)
-		}
-	} catch (error) {
-		// Mark socket as free
-		socketInUse = false
-		return Promise.reject(new Error(`Failed to fetch base template: ${error.message}`))
-	}
-}
-
-async function waitForSocketAvailability() {
-	return new Promise((resolve) => {
-		const checkInterval = setInterval(() => {
-			if (!socketInUse) {
-				clearInterval(checkInterval)
-				resolve()
-			}
-		}, 1000) // Check every second
-	})
-}
-/**
- *  download file from cloud and convert it into base64
- * @method
- * @name downloadAndConvertToBase64
- * @param {String} templateUrl - cloud path to download
- */
-async function downloadAndConvertToBase64(url) {
-	try {
-		// Wait if the socket is in use
-		if (socketInUse) {
-			console.log('Socket is in use. Waiting for 1 minute...')
-			await new Promise((resolve) => setTimeout(resolve, 60000)) // Wait for 1 minute
-		}
-
-		// Mark socket as in use
-		socketInUse = true
-
-		// Download the image file as a binary buffer
-		const response = await axios({
-			url,
-			method: 'GET',
-			responseType: 'arraybuffer', // Ensures we receive raw binary data
-			timeout: 120000,
-			headers: {
-				Connection: 'close', // Ensures the socket is closed after the request
-			},
-		})
-
-		// Convert the binary data to a Base64 string
-		const base64 = Buffer.from(response.data, 'binary').toString('base64')
-
-		// Create the Base64 Data URL
-		const base64DataUrl = `data:image/png;base64,${base64}`
-
-		// Mark socket as free
-		socketInUse = false
-
-		return base64DataUrl
-	} catch (error) {
-		console.error('Error downloading or converting file:', error.message)
-		// Mark socket as free
-		socketInUse = false
-		throw error
-	}
-}
-
-async function generatePresignedUrlInConsumption(url, body, headers) {
-	try {
-		const response = await axios.post(url, body, { headers, timeout: 6000 })
-		let result = { success: false }
-		if (response.status === 200) {
-			const files = response?.data?.result?.[common.CERTIFICATE]?.files
-
-			if (Array.isArray(files) && files.length > 0) {
-				result.file = files[0]?.payload?.sourcePath || null
-				result.url = files[0]?.url || null
-				result.success = true
-			} else {
-				console.error('Files array is missing or empty:', files)
-			}
-		} else {
-			console.error('Unexpected response status:', response.status)
-		}
-
-		return result
-	} catch (error) {
-		console.error('Error generating consumption presigned URL:', error.message)
-		throw error // Rethrow the error to be handled by the caller
-	}
-}
-
-async function uploadFile(dirPath, fileName, fileUploadUrl) {
-	try {
-		// Read the file data
-		const fileData = fs.readFileSync(path.join(dirPath, fileName))
-
-		const headers = {
-			'Content-Type': 'multipart/form-data',
-		}
-
-		// Perform the PUT request
-		const fileUploadToSignedUrl = await axios.put(fileUploadUrl, fileData, { headers })
-
-		// Check the response status
-		if (fileUploadToSignedUrl.status === 200) {
-			console.log('File uploaded successfully!')
-			console.log('Response status:', fileUploadToSignedUrl.status)
-		} else {
-			console.error('Unexpected response:', fileUploadToSignedUrl.status)
-		}
-	} catch (error) {
-		console.error('Error uploading file:', error.message)
-		if (error.response) {
-			console.error('Response status:', error.response.status)
-			console.error('Response data:', error.response.data)
-		}
-	}
-}
-/**
- * create svg template by editing base template.
- * @method
- * @name createSvg
- * @param {Object} certificateData - Certificate data for upload
- */
-
-async function createSvg(certificateData, loggedInUserId, userToken) {
-	return new Promise(async (resolve, reject) => {
-		try {
-			// fetch base template from cloud
-			let baseTemplate = await getBaseTemplate(certificateData?.base_template_url)
-			if (!baseTemplate.success) {
-				throw new Error('Base template download failed.')
-			}
-
-			// Load SVG template using Cheerio with XML mode
-			const $ = cheerio.load(baseTemplate.result, { xmlMode: true })
-
-			// set issuer name
-			const issuerNameTag = 'stateTitle'
-			const issuerNameElement = $(`#${issuerNameTag}`)
-			issuerNameElement.text(utils.escapeXml(certificateData.issuer))
-
-			// update signature
-			for (let index = 1; index <= certificateData.signature.no_of_signature; index++) {
-				const signatureNameTag = `signatureTitle${index}a`
-				const signatureDesignationTag = `signatureTitleDesignation${index}`
-				const signatureImgTag = `signatureImg${index}`
-				await waitForSocketAvailability() // check and wait for axios socket availability
-				const imageData = await downloadAndConvertToBase64(certificateData.signature[signatureImgTag])
-				const signatureNameElement = $(`#${signatureNameTag}`)
-				const signatureImgElement = $(`#${signatureImgTag}`)
-				signatureImgElement.attr('xlink:href', utils.escapeXml(imageData))
-				signatureNameElement.text(
-					`${utils.escapeXml(certificateData.signature[`signatureTitleName${index}`])} , ${utils.escapeXml(
-						certificateData.signature[signatureDesignationTag]
-					)}`
-				)
-			}
-
-			// update logos
-			for (let index = 1; index <= certificateData.logos.no_of_logos; index++) {
-				const logoTag = `stateLogo${index}`
-				await waitForSocketAvailability() // check and wait for axios socket availability
-				const imageData = await downloadAndConvertToBase64(certificateData.logos[logoTag])
-				const logoElement = $(`#${logoTag}`)
-				logoElement.attr('xlink:href', utils.escapeXml(imageData))
-			}
-
-			// updated svg
-			let updatedSvg = $.xml()
-
-			// replace quote escape charecters with "
-			updatedSvg = updatedSvg.replace(/&quot;/g, '"')
-
-			const uniqueId = utils.generateUniqueId() //generate a unique id for folder
-			let fileName = `${uniqueId}.svg` //create a unique file name
-			const mainPath = path.join(__dirname, `../temp/certificate/`) //temporary folder path for certificate template
-			let dirPath = path.join(mainPath, `${uniqueId}/`) //create a directory path
-			fs.mkdirSync(dirPath, { recursive: true }) //create directory
-			fs.writeFileSync(path.join(dirPath, fileName), updatedSvg, { encoding: 'utf8' }) //create file
-
-			// create a file upload payload
-			let payloadData = {
-				request: {
-					[common.CERTIFICATE]: {
-						files: [fileName],
-					},
-				},
-			}
-			// generate signed url
-			const headers = {
-				'X-auth-token': userToken.replace(/^bearer\s+/i, ''),
-			}
-			const getSignedUrl = await generatePresignedUrlInConsumption(
-				process.env.INTERFACE_SERVICE_HOST +
-					process.env.PROJECT_SERVICE_BASE_URL +
-					process.env.CONSUMPTION_SERVICE_PRESIGNED_URL,
-				payloadData,
-				headers
-			)
-			if (!getSignedUrl.success) {
-				throw new Error('FAILED_TO_GENERATE_SIGNED_URL')
-			}
-
-			const fileUploadUrl = getSignedUrl.url
-			let uploadedFilePath = getSignedUrl.file
-			await uploadFile(dirPath, fileName, fileUploadUrl)
-			// delete folder after upload
-			await deleteFolderRecursive(path.join(mainPath, uniqueId))
-
-			resolve({
-				message: 'Template edited successfully',
-				filePath: uploadedFilePath,
-			})
-		} catch (error) {
-			reject(error)
-		}
-	})
-}
-
-/**
- * function to recursively delete folder after upload
- * @method
- * @name deleteFolderRecursive
- * @param {String} folderPath - folder path to delete
- */
-async function deleteFolderRecursive(folderPath) {
-	// Check if the folder exists
-	if (fs.existsSync(folderPath)) {
-		// Get all files and subdirectories in the folder
-		fs.readdirSync(folderPath).forEach((file) => {
-			const currentPath = path.join(folderPath, file)
-
-			// If the item is a directory, recursively delete its contents
-			if (fs.lstatSync(currentPath).isDirectory()) {
-				deleteFolderRecursive(currentPath)
-			} else {
-				// Otherwise, delete the file
-				fs.unlinkSync(currentPath)
-			}
-		})
-		// Delete the empty folder
-		fs.rmdirSync(folderPath)
-		console.log(`Folder and its contents deleted: ${folderPath}`)
-	} else {
-		console.log('Folder does not exist:', folderPath)
-	}
-}
-/**
  * Check and Insert certificate base template
  * @method
  * @name checkCertificateBaseTemplate
@@ -907,7 +628,7 @@ async function insertCertificateTemplate(
 	orgCode,
 	tenantCode
 ) {
-	const svgTemplateCreation = await createSvg(certificateData, loggedInUserId, userToken)
+	const svgTemplateCreation = await certificateHelper.createSvg(certificateData, loggedInUserId, userToken)
 	const baseTemplate = await checkCertificateBaseTemplate(certificateData, orgCode, tenantCode)
 	const certificateDocument = {
 		status: common.STATUS_ACTIVE.toLowerCase(),
@@ -1755,11 +1476,12 @@ const publishProgram = function async(programData) {
 			//create user and program mapping
 			const viewerIds = rolloutDetails.viewers.map((viewer) => viewer?.id || viewer)
 			if (programId && viewerIds.length > 0) {
-				let createMappingResponse = await createOrUpdateUserProgramMapping(
+				let createMappingResponse = await userMappingHelper.createOrUpdateUserProgramMapping(
 					viewerIds,
 					programId,
 					programData.organization_code,
 					programData.tenant_code,
+					projectsMongoConnection,
 					loggedInUserId
 				)
 				console.log('User Program Mapping Response : ', createMappingResponse)
@@ -1787,103 +1509,6 @@ const publishProgram = function async(programData) {
 			)
 
 			return resolve(result)
-		}
-	})
-}
-
-/**
- * Maps users to a program using interfaceRequests.
- * @param {Array} viewers - Array of user IDs.
- * @param {String|ObjectId} programId - Program ID.
- * @param {String} tenantCode - Tenant code.
- * @param {String} orgCode - Organization code.
- * @param {String|null} userId - User ID performing the mapping.
- * @returns {Promise<Boolean>}
- */
-async function createOrUpdateUserProgramMapping(viewers, programId, orgCode, tenantCode, userId = null) {
-	return new Promise(async (resolve, reject) => {
-		try {
-			const roles = (process.env.DEFAULT_PROGRAM_MANAGERS || '')
-				.split(',')
-				.map((role) => role.trim())
-				.filter(Boolean)
-			if (roles.length === 0) {
-				throw new Error('No roles defined in DEFAULT_PROGRAM_MANAGERS environment variable')
-			}
-
-			const userProgramCollection = projectsMongoConnection.collection(COLLECTIONS_MAP.get('USER_EXTENSIONS'))
-			// Fetch all userExtensions for viewers
-			const userExtensions = await userProgramCollection.find({ userId: { $in: viewers } }).toArray()
-
-			// Find all userExtensions mapped to this program
-			const mappedUserExtensions = await userProgramCollection
-				.find({
-					'programRoleMapping.programId': programId,
-				})
-				.toArray()
-
-			// Users already mapped to this program
-			const alreadyMappedUserIds = mappedUserExtensions.map((userExt) => userExt.userId)
-
-			// Users in viewers but not mapped to program (need append)
-			const toAppend = viewers.filter((userId) => {
-				const ext = userExtensions.find((userExt) => userExt.userId === userId)
-				// If userExtension not present, need append
-				if (!ext) return true
-				// If userExtension present but programId not present, need append
-				const hasProgram = ext.programRoleMapping?.some((prm) => String(prm.programId) === String(programId))
-				return !hasProgram
-			})
-
-			// Users mapped to program but not in viewers (need remove)
-			// For each user mapped to the program but not present in viewers, prepare a remove operation
-			const toRemove = alreadyMappedUserIds.filter((userId) => !viewers.includes(userId))
-
-			// Prepare data for API call
-			const requestBody = []
-			// Only append if userExtension not present or programId not present in userExtension
-			programId = programId.toString()
-			if (toAppend.length > 0) {
-				for (const userId of toAppend) {
-					requestBody.push({
-						userId,
-						programId,
-						operation: common.OPERATION_APPEND,
-						roles: roles,
-					})
-				}
-			}
-
-			// Only remove if userExtension and programId present in userEx
-			// For each user mapped to the program but not present in viewers, prepare a remove operation
-			if (toRemove.length > 0) {
-				for (const userId of toRemove) {
-					requestBody.push({
-						userId,
-						programId,
-						operation: common.OPERATION_REMOVE,
-						roles: roles,
-					})
-				}
-			}
-
-			if (requestBody.length === 0) {
-				console.log('No user mapping changes required for program:', programId)
-				return resolve(true)
-			}
-			// Call the consumption service to update mappings
-			let userMappingResponse = await interfaceRequests.mapUserAndProgram(
-				requestBody,
-				orgCode,
-				tenantCode,
-				userId
-			)
-
-			console.log('Successfully updated user extensions for program:', programId)
-			return resolve(true)
-		} catch (error) {
-			console.error('Error in createOrUpdateUserProgramMapping:', error)
-			return reject(error)
 		}
 	})
 }
